WIP junkfood dropdown menu
