+++ src/main/java/dat/backend/control/Junkfood.java
@@ -1,12 +1,14 @@
 package dat.backend.control;
 
 import dat.backend.model.config.ApplicationStart;
+import dat.backend.model.entities.FoodAndDrinks;
 import dat.backend.model.persistence.ConnectionPool;
 
 import javax.servlet.*;
 import javax.servlet.http.*;
 import javax.servlet.annotation.*;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 @WebServlet(name = "Junkfood", value = "/Junkfood")
@@ -14,30 +16,33 @@ public class Junkfood extends HttpServlet {
 
     private ConnectionPool connectionPool;
 
-    private String food;
+    ArrayList<FoodAndDrinks> foodList = new ArrayList<>();
 
-    private int kcal;
+    final static FoodAndDrinks[] foodObjects = {
+            new FoodAndDrinks("Pizza", 600), //0
+            new FoodAndDrinks("Burger",700) //1
+    };
 
-    final static List<FoodAndDrinks> foodObjects = (new FoodAndDrinks("Pizza", 600));
 
-    public FoodAndDrinks(String food, int kcal) {
-        this.food = food;
-        this.kcal = kcal;
-    }
     @Override
     public void init() throws ServletException
     {
         this.connectionPool = ApplicationStart.getConnectionPool();
+
     }
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
+        //When the site is sent to client/browser then it will populate 'foodsanddrinks' object with the constant array foodObjects
+        request.setAttribute("foodsanddrinks", foodObjects);
         response.sendRedirect("junkfood.jsp");
+
+
     }
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
+        int selectedObejctIndex = Integer.parseInt(request.getParameter("junk"));
+        FoodAndDrinks selectedObject = foodObjects[selectedObejctIndex];
     }
 }
