+++ src/main/java/dat/backend/model/entities/FoodAndDrinks.java
@@ -1,0 +1,19 @@
+package dat.backend.model.entities;
+
+public class FoodAndDrinks {
+    private String food;
+    private int kcal;
+
+    public FoodAndDrinks(String food, int kcal) {
+        this.food = food;
+        this.kcal = kcal;
+    }
+
+    public String getFood() {
+        return food;
+    }
+
+    public int getKcal() {
+        return kcal;
+    }
+}
